+++ .circleci/config.templ.yml
@@ -548,6 +548,12 @@ jobs:
           pattern: "ci_visibility"
           trace_agent_url: ""
 
+  llmobs:
+    <<: *contrib_job_small
+    steps:
+      - run_test:
+          pattern: "llmobs"
+
   subprocess:
     <<: *contrib_job_small
     steps:
