--- ddtrace/internal/llmobs/__init__.py
@@ -1,4 +1,0 @@
-from .writer import LLMObsWriter
-
-
-__all__ = ["LLMObsWriter"]
