+++ ddtrace/llmobs/_llmobs.py
@@ -8,7 +8,7 @@ from ddtrace import config
 from ddtrace._trace.processor import TraceProcessor
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import atexit
-from ddtrace.internal.llmobs.writer import LLMObsWriter
+from ddtrace.llmobs._writer import LLMObsWriter
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.service import Service
