+++ tests/.suitespec.json
@@ -96,6 +96,9 @@
             "ddtrace/internal/ci_visibility/*",
             "ddtrace/ext/test.py"
         ],
+        "llmobs": [
+            "ddtrace/internal/llmobs/*"
+        ],
         "sourcecode": [
             "ddtrace/sourcecode/*"
         ],
@@ -503,6 +506,13 @@
             "@pytest",
             "tests/ci_visibility/*"
         ],
+        "llmobs": [
+            "@bootstrap",
+            "@core",
+            "@tracing",
+            "@llmobs",
+            "tests/llmobs/*"
+        ],
         "sourcecode": [
             "@bootstrap",
             "@core",
@@ -524,6 +534,7 @@
             "@contrib",
             "@openai",
             "@requests",
+            "@llmobs",
             "tests/contrib/openai/*",
             "tests/snapshots/tests.contrib.openai.*"
         ],
@@ -1214,6 +1225,7 @@
             "@contrib",
             "@langchain",
             "@requests",
+            "@llmobs",
             "tests/contrib/langchain/*",
             "tests/snapshots/tests.contrib.{suite}.*"
         ]
