+++ tests/contrib/botocore/test_bedrock.py
@@ -146,7 +146,7 @@ def bedrock_client(boto3, request_vcr):
 
 @pytest.fixture
 def mock_llmobs_writer():
-    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.LLMObsWriter")
+    patcher = mock.patch("ddtrace.llmobs._writer.LLMObsWriter")
     LLMObsWriterMock = patcher.start()
     m = mock.MagicMock()
     LLMObsWriterMock.return_value = m
