+++ tests/contrib/openai/conftest.py
@@ -115,7 +115,7 @@ def mock_logs(scope="session"):
     Note that this fixture must be ordered BEFORE mock_tracer as it needs to patch the log writer
     before it is instantiated.
     """
-    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.V2LogWriter")
+    patcher = mock.patch("ddtrace.llmobs._integrations.base.V2LogWriter")
     V2LogWriterMock = patcher.start()
     m = mock.MagicMock()
     V2LogWriterMock.return_value = m
@@ -125,7 +125,7 @@ def mock_logs(scope="session"):
 
 @pytest.fixture
 def mock_llmobs_writer(scope="session"):
-    patcher = mock.patch("ddtrace.internal.llmobs.integrations.base.LLMObsWriter")
+    patcher = mock.patch("ddtrace.llmobs._integrations.base.LLMObsWriter")
     LLMObsWriterMock = patcher.start()
     m = mock.MagicMock()
     LLMObsWriterMock.return_value = m
