+++ tests/llmobs/llmobs_cassettes/tests.internal.test_llmobs.test_send_chat_completion.yaml
