+++ tests/llmobs/llmobs_cassettes/tests.internal.test_llmobs.test_send_completion_bad_api_key.yaml
