+++ tests/llmobs/llmobs_cassettes/tests.internal.test_llmobs.test_send_completion_bad_app_key.yaml
